pyproject.toml
[project]
name = "geoguesser"
version = "0.1.0"
authors = [
  {name = "Julian Steinacher", email = "julian.steinacher@students.fhnw.ch"},
  {name = "Lukas Gysin", email = "lukas.gysin@students.fhnw.ch"},
]
dependencies = [
    "Pillow",
    "torch",
]

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
